Refactor DXF writer/reader, add decimal precision support
Refactored DXF stream writer and reader into new namespaces for better separation. Enhanced DxfAsciiWriter with configurable decimal precision for doubles, wired through DxfWriterConfiguration. Improved header variable management and documentation. Updated interfaces and base classes for clarity and consistency. Cleaned up namespaces and redundant code.
