+++ src/ACadSharp/IO/DXF/DxfStreamReader/DxfStreamReaderBase.cs
@@ -1,4 +1,5 @@
 ﻿using ACadSharp.Exceptions;
+using ACadSharp.IO.DXF.DxfStreamReader;
 using CSMath;
 using System;
 using System.IO;
