+++ src/ACadSharp/IO/DXF/DxfStreamWriter/DxfHeaderSectionWriter.cs
@@ -1,4 +1,5 @@
 ﻿using ACadSharp.Header;
+using ACadSharp.IO.DXF.DxfStreamWriter;
 using System.Collections.Generic;
 using System.Linq;
