+++ src/ACadSharp/IO/DXF/DxfStreamWriter/DxfTablesSectionWriter.cs
@@ -1,4 +1,5 @@
-﻿using ACadSharp.Tables;
+﻿using ACadSharp.IO.DXF.DxfStreamWriter;
+using ACadSharp.Tables;
 using ACadSharp.Tables.Collections;
 using CSMath;
 using System;
