+++ src/ACadSharp/IO/DxfWriterConfiguration.cs
@@ -3,132 +3,137 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace ACadSharp.IO
+namespace ACadSharp.IO;
+
+/// <summary>
+/// Configuration for writing DWG files.
+/// </summary>
+public class DxfWriterConfiguration : CadWriterConfiguration
 {
 	/// <summary>
-	/// Configuration for writing DWG files.
+	/// Gets or sets the double precision for writing the DXF file. 
+	/// If set to null, the default "R" format precision will be used.
 	/// </summary>
-	public class DxfWriterConfiguration : CadWriterConfiguration
+	public int? DecimalPrecision { get; set; } = null;
+
+	/// <summary>
+	/// Header variables to write in the dxf file
+	/// </summary>
+	public IEnumerable<string> HeaderVariables { get { return this._headerVariables.AsEnumerable(); } }
+
+	/// <summary>
+	/// Flag to write all the header variables in the dxf file
+	/// </summary>
+	/// <remarks>
+	/// Some variables are version sensitive, writing all the variables may cause a crash in other applications, to write an specific variable add it to the <see cref="AddHeaderVariable"/> set
+	/// </remarks>
+	/// <value>
+	/// false
+	/// </value>
+	public bool WriteAllHeaderVariables { get; set; } = false;
+
+	/// <summary>
+	/// Write the Dxf optional values.
+	/// </summary>
+	/// <remarks>
+	/// The values are marked with the attribute flag <see cref="DxfReferenceType.Optional"/>.
+	/// </remarks>
+	/// <value>
+	/// default: true
+	/// </value>
+	public bool WriteOptionalValues { get; set; } = true;
+
+	/// <summary>
+	/// Variables that must be writen in a dxf file
+	/// </summary>
+	public static readonly string[] Variables = new string[]
 	{
-		/// <summary>
-		/// Variables that must be writen in a dxf file
-		/// </summary>
-		public static readonly string[] Variables = new string[]
+		"$ACADVER",
+		"$DWGCODEPAGE",
+		"$LASTSAVEDBY",
+		"$HANDSEED",
+		"$ANGBASE",
+		"$ANGDIR",
+		"$ATTMODE",
+		"$AUNITS",
+		"$AUPREC",
+		"$CECOLOR",
+		"$CELTSCALE",
+		"$CELTYPE",
+		"$CELWEIGHT",
+		"$CLAYER",
+		"$CMLJUST",
+		"$CMLSCALE",
+		"$CMLSTYLE",
+		"$DIMSTYLE",
+		"$TEXTSIZE",
+		"$TEXTSTYLE",
+		"$LUNITS",
+		"$LUPREC",
+		"$MIRRTEXT",
+		"$EXTNAMES",
+		"$INSBASE",
+		"$INSUNITS",
+		"$MEASUREMENT",
+		"$LTSCALE",
+		"$LWDISPLAY",
+		"$PDMODE",
+		"$PDSIZE",
+		"$PLINEGEN",
+		"$PSLTSCALE",
+		"$SPLINESEGS",
+		"$SURFU",
+		"$SURFV",
+		"$TDCREATE",
+		"$TDUCREATE",
+		"$TDUPDATE",
+		"$TDUUPDATE",
+		"$TDINDWG",
+	};
+
+	private HashSet<string> _headerVariables;
+
+	public DxfWriterConfiguration()
+	{
+		this._headerVariables = new HashSet<string>(Variables);
+	}
+
+	/// <summary>
+	/// Add a Header variable name to be added in the dxf document
+	/// </summary>
+	/// <remarks>
+	/// The name of the variable must exist in <see cref="CadHeader.GetHeaderMap"/>
+	/// </remarks>
+	/// <param name="name"></param>
+	/// <exception cref="ArgumentException"></exception>
+	public void AddHeaderVariable(string name)
+	{
+		Dictionary<string, CadSystemVariable> map = CadHeader.GetHeaderMap();
+
+		if (!map.ContainsKey(name))
 		{
-			"$ACADVER",
-			"$DWGCODEPAGE",
-			"$LASTSAVEDBY",
-			"$HANDSEED",
-			"$ANGBASE",
-			"$ANGDIR",
-			"$ATTMODE",
-			"$AUNITS",
-			"$AUPREC",
-			"$CECOLOR",
-			"$CELTSCALE",
-			"$CELTYPE",
-			"$CELWEIGHT",
-			"$CLAYER",
-			"$CMLJUST",
-			"$CMLSCALE",
-			"$CMLSTYLE",
-			"$DIMSTYLE",
-			"$TEXTSIZE",
-			"$TEXTSTYLE",
-			"$LUNITS",
-			"$LUPREC",
-			"$MIRRTEXT",
-			"$EXTNAMES",
-			"$INSBASE",
-			"$INSUNITS",
-			"$MEASUREMENT",
-			"$LTSCALE",
-			"$LWDISPLAY",
-			"$PDMODE",
-			"$PDSIZE",
-			"$PLINEGEN",
-			"$PSLTSCALE",
-			"$SPLINESEGS",
-			"$SURFU",
-			"$SURFV",
-			"$TDCREATE",
-			"$TDUCREATE",
-			"$TDUPDATE",
-			"$TDUUPDATE",
-			"$TDINDWG",
-		};
-
-		/// <summary>
-		/// Flag to write all the header variables in the dxf file
-		/// </summary>
-		/// <remarks>
-		/// Some variables are version sensitive, writing all the variables may cause a crash in other applications, to write an specific variable add it to the <see cref="AddHeaderVariable"/> set
-		/// </remarks>
-		/// <value>
-		/// false
-		/// </value>
-		public bool WriteAllHeaderVariables { get; set; } = false;
-
-		/// <summary>
-		/// Write the Dxf optional values.
-		/// </summary>
-		/// <remarks>
-		/// The values are marked with the attribute flag <see cref="DxfReferenceType.Optional"/>.
-		/// </remarks>
-		/// <value>
-		/// default: true
-		/// </value>
-		public bool WriteOptionalValues { get; set; } = true;
-
-		/// <summary>
-		/// Header variables to write in the dxf file
-		/// </summary>
-		public IEnumerable<string> HeaderVariables { get { return this._headerVariables.AsEnumerable(); } }
-
-		private HashSet<string> _headerVariables;
-
-		public DxfWriterConfiguration()
-		{
-			this._headerVariables = new HashSet<string>(Variables);
+			throw new ArgumentException($"The variable {name} does not exist in the header", nameof(name));
 		}
 
-		/// <summary>
-		/// Add a Header variable name to be added in the dxf document
-		/// </summary>
-		/// <remarks>
-		/// The name of the variable must exist in <see cref="CadHeader.GetHeaderMap"/>
-		/// </remarks>
-		/// <param name="name"></param>
-		/// <exception cref="ArgumentException"></exception>
-		public void AddHeaderVariable(string name)
+		this._headerVariables.Add(name);
+	}
+
+	/// <summary>
+	/// Remove a Header variable name so is not added in the dxf document
+	/// </summary>
+	/// <remarks>
+	/// The name cannot be a in the <see cref="Variables"/> list
+	/// </remarks>
+	/// <param name="name"></param>
+	/// <returns></returns>
+	/// <exception cref="ArgumentException"></exception>
+	public bool RemoveHeaderVariable(string name)
+	{
+		if (Variables.Select(v => v.ToLowerInvariant()).Contains(name.ToLowerInvariant()))
 		{
-			Dictionary<string, CadSystemVariable> map = CadHeader.GetHeaderMap();
-
-			if (!map.ContainsKey(name))
-			{
-				throw new ArgumentException($"The variable {name} does not exist in the header", nameof(name));
-			}
-
-			this._headerVariables.Add(name);
+			throw new ArgumentException($"The variable {name} cannot be removed from the set", nameof(name));
 		}
 
-		/// <summary>
-		/// Remove a Header variable name so is not added in the dxf document
-		/// </summary>
-		/// <remarks>
-		/// The name cannot be a in the <see cref="Variables"/> list
-		/// </remarks>
-		/// <param name="name"></param>
-		/// <returns></returns>
-		/// <exception cref="ArgumentException"></exception>
-		public bool RemoveHeaderVariable(string name)
-		{
-			if (Variables.Select(v => v.ToLowerInvariant()).Contains(name.ToLowerInvariant()))
-			{
-				throw new ArgumentException($"The variable {name} cannot be removed from the set", nameof(name));
-			}
-
-			return this._headerVariables.Remove(name);
-		}
+		return this._headerVariables.Remove(name);
 	}
 }
